+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - YYYY-MM-DD
 
+No items.
+
+## [0.3.0] - 2026-06-24
+
 ### Fixed
 
 - `ArrDeque::clone` (NG: bit copy, not clone)
@@ -32,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/nossie531/arr_deque/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/nossie531/arr_deque/compare/v0.3.0...HEAD
+[0.2.0]: https://github.com/nossie531/arr_deque/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/nossie531/arr_deque/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/nossie531/arr_deque/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/nossie531/arr_deque/compare/v0.1.0...v0.1.1